requirements.txt
pandas
Pillow
loguru
fastapi
uvicorn[standard]
xlsxwriter
openpyxl
python-multipart